Renamve to ParticipantsRegularGrid
