+++ docusaurus/docs/Android/04-ui-components/04-call/06-screen-share-content.mdx
@@ -24,7 +24,7 @@ if (screenSharingSession != null) {
     )
 } else {
    // render a regular video call screen
-   RegularCallVideoRenderer(
+   ParticipantsRegularGrid(
        call = call,
        modifier = modifier,
        onRender = onRender,
